requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
sqlalchemy>=1.4,<1.5  # Use SQLAlchemy 1.4.x for compatibility with databases
databases[postgresql]==0.6.0
psycopg2-binary==2.9.8
pytest==7.4.0
pytest-asyncio==0.21.0
httpx==0.24.0
black==23.7.0
flake8==6.1.0
isort==5.12.0
python-dotenv==1.0.0
